feat: scale to $100/day — 6 assets, compounding, Monte Carlo, Pro TUI
Multi-Asset Expansion:
  - Added DOGE + MATIC (6 assets total)
  - Binance WS stream expanded for 6 symbols
  - Per-asset goroutine-per-market with LockOSThread

SharedMemoryHub (cross-asset correlation):
  - Lock-free ring buffer (256 entries per asset)
  - SMA/deviation computation in pure integer arithmetic
  - Pearson correlation between any 2 assets (~200ns for N=60)

DailyTracker (compounding + money management):
  - Compound stake +5% per win, reset on loss
  - 20% drawdown from daily peak → Safety Mode (30-min cooldown)
  - Atomic win/loss tracking per asset
  - Daily progress toward $100 target

Monte Carlo Simulator (--analyze mode):
  - 10,000 simulated trading days
  - Percentile distribution (P5/P25/P50/P75/P90/P95/P99)
  - Visual histogram in terminal
  - Win rate estimation from historical price data

Panel TUI Pro:
  - Per-asset table: price, SMA(60), deviation, W/L ratio, BTC correlation
  - Daily progress bar toward $100 target
  - Compounding stake display
  - Safety mode indicator with ANSI colors

Usage:
  ./sniper             # normal trading mode
  ./sniper --analyze   # Monte Carlo profit projection

+++ go/internal/models/fixedpoint.go
@@ -63,14 +63,14 @@ func (fp FixedPoint) Float64() float64 {
 //
 //go:nosplit
 func (fp FixedPoint) Cents() int64 {
-	return int64(fp / (FPScale / 100))
+	return int64(uint64(fp) / (FPScale / 100))
 }
 
 // Dollars returns the integer dollar amount (truncated).
 //
 //go:nosplit
 func (fp FixedPoint) Dollars() int64 {
-	return int64(fp / FPScale)
+	return int64(uint64(fp) / FPScale)
 }
 
 // Add returns fp + other. Zero allocation.
